Add screenshots of the app
The diffs for lines 6-9 of the README file may
look strange if your display wraps text to the
next line instead of allowing horizontal
scrolling.
Essentially, Prettier formats the table cells to
be as wide as the widest cell in the same column, so
the image links make each cell very wide and cause
the words to wrap.
Enable horizontal scrolling to see the actual
arrangement of the cells.