Test if Sidebar's width is set by child Box
This again refactors the test to only check for the presence of
the width prop. The previous test for a component's display
name was not ideal because it required extracting the
functionality within the Box into a separate component.
Also, the value of the prop is based on the 'size' context.